Match func_002d8a60 shop list row drawer from C
Rewrite the guarded draft in file idiom over a file-local ShopListWork
overlay (list/slide/select/scroll fields), with scoped opt_loop_invariants
on. func_002e2830 returns int: retail callers re-extend its result, and the
y_list definition stays byte-identical.

+++ src/Yajima/y_list.c
@@ -163,7 +163,7 @@ s16 func_002e2740(s32 arg0) {
 }
 
 // FUN_002E2830
-s16 func_002e2830(u8 *arg0, s32 arg1) {
+s32 func_002e2830(u8 *arg0, s32 arg1) {
     u8 *p = *(u8 **)(arg0 + 0x38);
 
     if (*(s16 *)(p + 2) < arg1) {
